Adds velero plugin for vsphere rock and sanity test
Adds the rock for the mentioned velero plugin.

Note that we're only building for amd64, since the plugin depends on
VDDX libraries from VMWare, and there doesn't seem to be support for arm64.

Added a sleep to the command so it doesn't exit immediately. This way,
Pebble respects the on-success / on-failure rules.